requirements.txt
tensorflow==2.16.1
fastapi==0.110.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
pillow==10.3.0
numpy==1.26.4
streamlit==1.33.0
requests==2.31.0
matplotlib==3.8.4
pandas==2.2.2
locust==2.24.0
scikit-learn>=1.3.0
